railway-crossing/services/services/requirements.txt
uvicorn[standard]
fastapi[all]
protobuf
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp